backend/requirements.txt
fastapi
uvicorn[standard]
tinydb
bcrypt
python-jose[cryptography]
requests
python-dotenv